Switch to new ZSH config
This is probably the biggest change I've made in a long while. It
switches to use antibody to manage zsh plugins rather than the system
I had before. As a result opening new shells is much faster than before.
The prompt has also changed because I'm too lazy to port over the old
prompt to a plugin and this new one is probably more performant anyway.

Along with the change, I've switched to use asdf to manage all the
different versions of tools I need. Python is slightly harder because
there's no virtualenv plugin, but everything is simpler without having
to use chruby, pyenv, and nvm and it makes it easy to use new languages
as well.